+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Added statement duration to logging
 - Added logging for `dbxUpdate` transaction
+- Added `variables` to `dbxConnect`
 
 # dbx 0.2.1